backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain-google-genai==2.0.7
langchain==0.3.11
pdfplumber
pydantic
langchain-core==0.3.25
langchain-text-splitters==0.3.2
langsmith==0.2.3
google-generativeai==0.8.3
google-ai-generativelanguage==0.6.10
google-api-python-client==2.155.0
httplib2==0.22.0
uritemplate==4.1.1
langchain-community
pypdf
